+++ semcore/data-table/__tests__/data-table-accordion.browser-test.tsx
@@ -1,6 +1,7 @@
 import type { Page } from '@playwright/test';
 import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 import { expect, test } from '@semcore/testing-utils/playwright';
+import { loadPage } from '@tools/testing-utils/shared/helpers';
 
 const checkStyles = async (element: any, styles: Record<string, string>) => {
   for (const [property, value] of Object.entries(styles) as [string, string][]) {
@@ -1320,5 +1321,22 @@ test.describe('Accordion in table', () => {
         'background-color': 'rgb(196, 229, 254)',
       });
     }
+  });
+
+  test('Verify accordion after skeleton in table cell', async ({ page }) => {
+    await loadPage(page, 'stories/components/data-table/tests/examples/accordion-tests/accordion-in-table-in-table.tsx', 'en', {
+      withSkeletonsAndAsyncDataLoading: true,
+    });
+
+    const row = locators.row(page, 4);
+    const cell = row.locator('[role="gridcell"][aria-colindex="4"]');
+    const accordionCellButton = cell.locator('button');
+    const collapse = locators.collapse(page);
+
+    await accordionCellButton.waitFor({ state: 'visible' });
+
+    await accordionCellButton.click();
+
+    await expect(collapse).toBeVisible({ timeout: 500 });
   });
 });