+++ semcore/date-picker/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.2.1] - 2025-10-24
+## [16.2.2] - 2025-11-10
 
 ### Fixed
 
 - Space key interactions in Date Picker (unable to switch months, apply Today, or keep selected date)
+
+## [16.2.1] - 2025-10-29
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/base-trigger` [16.4.2 ~> 16.4.3], `@semcore/button` [16.0.10 ~> 16.0.11], `@semcore/icon` [16.5.1 ~> 16.7.2], `@semcore/checkbox` [16.2.0 ~> 16.2.1], `@semcore/tooltip` [16.0.10 ~> 16.0.11], `@semcore/base-components` [16.4.0 ~> 16.4.1]).
 
 ## [16.2.0] - 2025-10-03